Implemented new artwork system, added new slide state, removed pickups in favor of new ideas
Came back to the project after some Godot bootcamp and I've decided to re-tool most of it. I finally deciphered most of what was going on in the state machine so I can start working on this again. Along the way I discovered resource files and how to implement atlases so I will be doing so going forward to keep artwork data and other things more organized.